requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autopep8==1.5.7
boto==2.49.0
boto3==1.20.8
botocore==1.23.8
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
decorator==5.1.0
Django==3.2.6
django-crispy-forms==1.13.0
django-filter==2.4.0
django-storages==1.12.3
django-widget-tweaks==1.4.8
folium==0.12.1
frozenlist==1.2.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geoip2==4.4.0
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==5.2.0
numpy==1.21.3
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
yarl==1.7.2
